requirements.txt
alembic==1.4.3
click==7.1.2
coverage==5.3
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-jsonapi==0.23.2
nose==1.3.7
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
